refactor: convert NotFound component to Typescript (only file name extension change needed)
